requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.4.0
django-filter==24.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-ratelimit==4.1.0
drf-spectacular==0.27.2
qrcode==8.2
django-oauth-toolkit==3.0.1
django-recaptcha==4.1.0
